+++ fns-hl7-pipeline/fn-redactor/src/main/kotlin/gov/cdc/dex/hl7/model/RedactorReport.kt
@@ -1,0 +1,8 @@
+package gov.cdc.dex.hl7.model
+
+import gov.cdc.hl7.RedactInfo
+
+data class RedactorReport(val entries: List<RedactInfo>) {
+    val status = "SUCCESS"
+
+}