+++ svc-cloud-transport/src/main/kotlin/gov/cdc/dex/transport/service/Application.kt
@@ -1,0 +1,15 @@
+package gov.cdc.dex.transport.service
+
+import io.micronaut.runtime.Micronaut
+
+object Application {
+
+    @JvmStatic
+    fun main(args: Array<String>) {
+//        Micronaut.run(Application.javaClass)
+        Micronaut.build()
+            .mainClass(Application.javaClass)
+            .environmentPropertySource(true)
+            .start()
+    }
+}