+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.0.14] - 2023-03-08
+  - Set up TST environment and pipelines
+  - Develop transport service to upload files with Metadata (pending deployment)
+  - Designed and Develop Transport metadata to identify messages upon upload
+  - Flaggged cardinality over as warnings
+  - Added vocabulary check for OBR-31/GenV2 messages
+  - Added COVID ELR profiles for structure validation
+  - Bug fixes and Enhancements
+    - Fixed NK1 Redaction paths
+    - Routing unknwon messages properly
+    - Fixed MMG-Based and MMG-In-SQL transform to support non-GenV2 messages
+  
+  
+
 ### [0.0.13] - 2023-02-22
   - Created new Function to Redact messages
   - Adding MMG for Varicella and TB Messages